added a download files method
Will download files from the specified location within s3.  It will auto
add the extension ".tran" when downloading, when the download is
finished it will remove this extension.  This allows for other
processors to be constantly eating files (without the .tran extension)
without worrying about eating a file that isn't finished downloading