bug fix and added non-generic bulkinserter and maps
fixed bug (typo) which caused passing in a connection string name rather
than the connection string to not work with the bulk inserter.

Also added MapDirectly() which maps props to column names 1-1 and this
is the default if nothing is specified.

Non generic versions of BulkInserter and BulkInserterColumnMappings,
allowing you to instead pass in the type of the object to be inserted